Add fit=fill feature for resize
When resizing an image which is smaller than the dimensions you want to
resize it to in one or both directions, the fit=fill option allows you
to specify a fill color (when combined with the Background manipulator)
to fill the extra canvas space.

This is accomplished using a query like `?w=300&fit=fill&bg=000` to resize
the image to 300 pixels wide while filling any new canvas space with
black.