Makes temp dir configurable
These changes make it posible to set the temp dir used by
the glide server to store images for processing.

Since this must be a local path we check if the path exists and is
writable by php. Otherwise an exception is thrown.

By default the tempDir property is set to sys_get_temp_dir() so no
changes are required for existing systems.